# speed up integration tests by using the asserting the headers and body of a single response