--- a/test/integration/files.test.js
+++ b/test/integration/files.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const request = require('supertest');
+const {assert} = require('chai');
 
 describe('GET /files', function() {
 	this.timeout(20000);
@@ -10,18 +11,22 @@ describe('GET /files', function() {
 		const moduleName = 'o-test-component@1.0.13';
 		const pathName = 'readme.md';
 
-		beforeEach(function() {
-			this.request = request(this.app)
+		/**
+		 * @type {request.Response}
+		 */
+		let response;
+		before(async function () {
+			response = await request(this.app)
 				.get(`/files/${moduleName}/${pathName}`)
 				.set('Connection', 'close');
 		});
 
-		it('should respond with a 301 status', function(done) {
-			this.request.expect(301).end(done);
+		it('should respond with a 301 status', function() {
+			assert.equal(response.status, 301);
 		});
 
-		it('should respond with a v2 `Location` header', function(done) {
-			this.request.expect('Location', `${this.basepath}v2/files/${moduleName}/${pathName}`).end(done);
+		it('should respond with a v2 `Location` header', function() {
+			assert.deepEqual(response.headers['location'], `${this.basepath}v2/files/${moduleName}/${pathName}`);
 		});
 
 	});