--- a/test/integration/home.test.js
+++ b/test/integration/home.test.js
@@ -1,27 +1,32 @@
 'use strict';
 
 const request = require('supertest');
+const {assert} = require('chai');
 
 describe('GET /', function() {
 	this.timeout(20000);
 	this.slow(5000);
 
-	beforeEach(function() {
-		this.request = request(this.app)
+	/**
+		 * @type {request.Response}
+		 */
+	let response;
+	before(async function () {
+		response = await request(this.app)
 			.get('/')
 			.set('Connection', 'close');
 	});
 
-	it('should respond with a 301 status', function(done) {
-		this.request.expect(301).end(done);
+	it('should respond with a 301 status', function() {
+		assert.equal(response.status, 301);
 	});
 
-	it('should respond with a v2 `Location` header', function(done) {
-		this.request.expect('Location', `${this.basepath}v2/`).end(done);
+	it('should respond with a v2 `Location` header', function() {
+		assert.deepEqual(response.headers['location'], `${this.basepath}v2/`);
 	});
 
-	it('should respond with plaintext', function(done) {
-		this.request.expect('Content-Type', 'text/plain; charset=utf-8').end(done);
+	it('should respond with plaintext', function() {
+		assert.deepEqual(response.headers['content-type'], 'text/plain; charset=utf-8');
 	});
 
 });