--- a/test/integration/modules.test.js
+++ b/test/integration/modules.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const request = require('supertest');
+const {assert} = require('chai');
 
 describe('GET /modules', function() {
 	this.timeout(20000);
@@ -9,18 +10,22 @@ describe('GET /modules', function() {
 	describe('when a valid module is requested', function() {
 		const moduleName = 'o-test-component%401.0.16';
 
-		beforeEach(function() {
-			this.request = request(this.app)
+		/**
+		 * @type {request.Response}
+		 */
+		let response;
+		before(async function () {
+			response = await request(this.app)
 				.get(`/modules/${moduleName}`)
 				.set('Connection', 'close');
 		});
 
-		it('should respond with a 301 status', function(done) {
-			this.request.expect(301).end(done);
+		it('should respond with a 301 status', function() {
+			assert.equal(response.status, 301);
 		});
 
-		it('should respond with a v2 `Location` header', function(done) {
-			this.request.expect('Location', `${this.basepath}v2/modules/${moduleName}`).end(done);
+		it('should respond with a v2 `Location` header', function() {
+			assert.deepEqual(response.headers['location'], `${this.basepath}v2/modules/${moduleName}`);
 		});
 
 	});