our previous approach was to make a new request/response for each assertion, even if the request was for the exact same resource we had just requested.

Running the integration tests against a local application:
- Without this change: 4 minutes 5 seconds
- With this change: 1 minute 30 seconds.

Running the integration tests against the heroku application:
- Without this change: 8 minutes 43 seconds
- With this change: 3 minutes 26 seconds